CBQE-0: fix killing process of n1ql
Reviewed-on: http://review.couchbase.org/49595
Tested-by: buildbot <build@couchbase.com>

+++ pytests/tuqquery/tuq.py
@@ -89,8 +89,10 @@ class QueryTests(BaseTestCase):
                if len(o):
                    for cbq_engine in o[0]:
                        if cbq_engine.find('grep') == -1:
-                           pid = [item for item in cbq_engine.split(' ') if item][1]
-                           self.shell.execute_command("kill -9 %s" % pid)
+                           output = cbq_engine.split(' ')
+                           if len(output) > 1:
+                                pid = [item for item in output if item][1]
+                                self.shell.execute_command("kill -9 %s" % pid)
 
 
 ##############################################################################################
